added env conditional to dev info box and work on renaming student course components
