+++ frontend/public/styles/fonts.css
@@ -1,32 +1,32 @@
-@font-face{
-    font-family: 'BCSans';
-    font-style: normal;
-    src: url('BCSans-Regular.woff') format('woff'),
-    url('BCSans-Regular.woff2') format('woff2');
+@font-face {
+  font-family: "BCSans";
+  font-style: normal;
+  src: url("BCSans-Regular.woff") format("woff"),
+    url("BCSans-Regular.woff2") format("woff2");
 }
-@font-face{
-    font-family: 'BCSans';
-    font-style: italic;
-    src: url('BCSans-Italic.woff') format('woff'),
-    url('BCSans-Italic.woff2') format('woff2');
+@font-face {
+  font-family: "BCSans";
+  font-style: italic;
+  src: url("BCSans-Italic.woff") format("woff"),
+    url("BCSans-Italic.woff2") format("woff2");
 }
-@font-face{
-    font-family: 'BCSans';
-    font-weight: 700;
-    src: url('BCSans-Bold.woff') format('woff'),
-    url('BCSans-Bold.woff2') format('woff2');
+@font-face {
+  font-family: "BCSans";
+  font-weight: 700;
+  src: url("BCSans-Bold.woff") format("woff"),
+    url("BCSans-Bold.woff2") format("woff2");
 }
-@font-face{
-    font-family: 'BCSans';
-    font-style: italic;
-    font-weight: 700;
-    src: url('BCSans-BoldItalic.woff') format('woff'),
-    url('BCSans-BoldItalic.woff2') format('woff2');
+@font-face {
+  font-family: "BCSans";
+  font-style: italic;
+  font-weight: 700;
+  src: url("BCSans-BoldItalic.woff") format("woff"),
+    url("BCSans-BoldItalic.woff2") format("woff2");
 }
 
 .v-application {
-    font-family: 'BCSans', Verdana, Arial, sans-serif;
+  font-family: "BCSans", Verdana, Arial, sans-serif;
 }
-html{
-    font-family: 'BCSans', Verdana, Arial, sans-serif;
+html {
+  font-family: "BCSans", Verdana, Arial, sans-serif;
 }
