+++ frontend/src/components/StudentProfile/Forms/FormInputs/CourseDetailsInput.vue
@@ -1,0 +1,10 @@
+<template>
+  <v-row no-gutters></v-row>
+</template>
+
+<script>
+export default {
+  name: "CourseDetailsInput",
+  props: { courseId: String, sessionDate: String },
+};
+</script>
